requirements.txt
pyinstaller==4.7
pytest==6.2.5
pytest-mock==3.6.1
pillow==8.4.0
